tile_service/requirements.txt
requests==2.26.0
numpy==1.23
pandas==1.3.4
matplotlib==3.5.1
opencv-python==4.5.5.64
lxml==5.2.1
scikit-image==0.19.3
pyserial==3.5
crc==1.3.0
imageio==2.31.1
tifffile==2023.7.10
hypha-rpc
zarr==2.11.3
flask
python-dotenv
openai
blosc
aiohttp==3.10.1